+++ tests/testthat/test-get_ag_bulletin.R
@@ -1,10 +1,10 @@
 context("get_ag_bulletin")
 
-# Test that get_ag_bulletin returns a data frame with 20 columns ------------------
-test_that("get_ag_bulletin returns 20 columns", {
+# Test that get_ag_bulletin returns a data frame with 20 columns ---------------
+test_that("get_ag_bulletin returns 27 columns", {
   skip_on_cran()
   BOM_bulletin <- get_ag_bulletin(state = "QLD")
-  expect_equal(ncol(BOM_bulletin), 21)
+  expect_equal(ncol(BOM_bulletin), 27)
   expect_named(
     BOM_bulletin,
     c(
@@ -12,7 +12,16 @@ test_that("get_ag_bulletin returns 20 columns", {
       "obs_time_utc",
       "time_zone",
       "site",
-      "name",
+      "dist",
+      "station",
+      "start",
+      "end",
+      "state",
+      "lat",
+      "lon",
+      "elev",
+      "bar_ht",
+      "WMO",
       "r",
       "tn",
       "tx",
@@ -25,63 +34,59 @@ test_that("get_ag_bulletin returns 20 columns", {
       "t20",
       "t50",
       "t1m",
-      "wr",
-      "state",
-      "lat",
-      "lon"
+      "wr"
     )
   )
 })
 
-
-# Test that get_ag_bulletin returns the requested state bulletin ------------------
+# Test that get_ag_bulletin returns the requested state bulletin ---------------
 test_that("get_ag_bulletin returns the bulletin for ACT/NSW", {
   skip_on_cran()
-  BOM_bulletin <- as.data.frame(get_ag_bulletin(state = "NSW"))
-  expect_equal(BOM_bulletin[1, 19], "NSW")
+  BOM_bulletin <- get_ag_bulletin(state = "NSW")
+  expect_equal(BOM_bulletin[1, "state"], "NSW")
 })
 
 test_that("get_ag_bulletin returns the bulletin for NT", {
   skip_on_cran()
-  BOM_bulletin <- as.data.frame(get_ag_bulletin(state = "NT"))
-  expect_equal(BOM_bulletin[1, 19], "NT")
+  BOM_bulletin <- get_ag_bulletin(state = "NT")
+  expect_equal(BOM_bulletin[1, 9], "NT")
 })
 
 test_that("get_ag_bulletin returns the bulletin for QLD", {
   skip_on_cran()
-  BOM_bulletin <- as.data.frame(get_ag_bulletin(state = "QLD"))
-  expect_equal(BOM_bulletin[1, 19], "QLD")
+  BOM_bulletin <- get_ag_bulletin(state = "QLD")
+  expect_equal(BOM_bulletin[1, "state"], "QLD")
 })
 
 test_that("get_ag_bulletin returns the bulletin for SA", {
   skip_on_cran()
-  BOM_bulletin <- as.data.frame(get_ag_bulletin(state = "SA"))
-  expect_equal(BOM_bulletin[1, 19], "SA")
+  BOM_bulletin <- get_ag_bulletin(state = "SA")
+  expect_equal(BOM_bulletin[1, "state"], "SA")
 })
 
 test_that("get_ag_bulletin returns the bulletin for TAS", {
   skip_on_cran()
-  BOM_bulletin <- as.data.frame(get_ag_bulletin(state = "TAS"))
-  expect_equal(BOM_bulletin[1, 19], "TAS")
+  BOM_bulletin <- get_ag_bulletin(state = "TAS")
+  expect_equal(BOM_bulletin[1, "state"], "TAS")
 })
 
 test_that("get_ag_bulletin returns the bulletin for VIC", {
   skip_on_cran()
-  BOM_bulletin <- as.data.frame(get_ag_bulletin(state = "VIC"))
-  expect_equal(BOM_bulletin[1, 19], "VIC")
+  BOM_bulletin <- get_ag_bulletin(state = "VIC")
+  expect_equal(BOM_bulletin[1, "state"], "VIC")
 })
 
 test_that("get_ag_bulletin returns the bulletin for WA", {
   skip_on_cran()
-  BOM_bulletin <- as.data.frame(get_ag_bulletin(state = "WA"))
-  expect_equal(BOM_bulletin[1, 19], "WA")
+  BOM_bulletin <- get_ag_bulletin(state = "WA")
+  expect_equal(BOM_bulletin[1, "state"], "WA")
 })
 
 test_that("get_ag_bulletin returns the bulletin for AUS", {
   skip_on_cran()
-  BOM_bulletin <- as.data.frame(get_ag_bulletin(state = "AUS"))
-  expect_equal(unique(BOM_bulletin[, 19]),
-               c("NT", "NSW", "QLD", "SA", "TAS", "VIC", "WA"))
+  BOM_bulletin <- get_ag_bulletin(state = "AUS")
+  state <- na.omit(BOM_bulletin["state"])
+  expect_equal(nrow(unique(state)), 7)
 })
 
 # Test that .validate_state stops if the state recognised ----------------------