+++ tests/testthat/test-get_current_weather.R
@@ -8,8 +8,6 @@ test_that("Error handling", {
                  regexp = "Both station_name and latlon provided. Ignoring latlon")
   expect_error(get_current_weather(latlon = 33), regexp = "[Ll]ength")
   expect_error(get_current_weather(latlon = c("-33", "151")), regexp = "[Nn]umeric")
-  expect_warning(get_current_weather(latlon = c(0, 0)), regexp = "unlikely")
-  expect_warning(get_current_weather(latlon = c(-33, -151)), regexp = "unlikely")
 })
 
 test_that("Query of 'Melbourne Airport' returns data frame with correct station", {
@@ -26,7 +24,7 @@ test_that("Query of 'Melbourne Airport' returns time if cooked.", {
 })
 
 test_that("latlon: Query of c(-27, 149) returns Surat (QLD, between Roma and St George).", {
-  expect_message(get_current_weather(latlon = c(-27, 149)), regexp = "Surat")
+  expect_message(get_current_weather(latlon = c(-27, 149)), regexp = "SURAT")
   Surat <- get_current_weather(latlon = c(-27, 149), emit_latlon_msg = FALSE)
   expect_equal(unique(Surat$name), "Surat")
 })